requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
requests
beautifulsoup4
google-generativeai
pydantic>=2.5,<3.0
bcrypt
PyJWT
email-validator
